Fix build error due to mismatched types
Error was:

| literals.cpp: In function 'libp2p::common::Hash256 libp2p::common::operator""_hash256(const char*, size_t)':
| literals.cpp:19:36: error: no matching function for call to 'min(size_t&, long unsigned int)'
|    19 |     std::copy_n(c, std::min(s, 32ul), hash.rbegin());
|       |                                    ^

+++ src/common/literals.cpp
@@ -16,13 +16,13 @@
 namespace libp2p::common {
   libp2p::common::Hash256 operator""_hash256(const char *c, size_t s) {
     libp2p::common::Hash256 hash{};
-    std::copy_n(c, std::min(s, 32ul), hash.rbegin());
+    std::copy_n(c, std::min(s, static_cast<size_t>(32u)), hash.rbegin());
     return hash;
   }
 
   libp2p::common::Hash512 operator""_hash512(const char *c, size_t s) {
     libp2p::common::Hash512 hash{};
-    std::copy_n(c, std::min(s, 64ul), hash.rbegin());
+    std::copy_n(c, std::min(s, static_cast<size_t>(64u)), hash.rbegin());
     return hash;
   }
